backend/requirements.txt
flask
flask-cors
pymongo
bcrypt
resend
python-dotenv
flask-limiter
flask-socketio
